real_time_stt/requirements.txt
faster-whisper>=0.4.1
sounddevice>=0.4.6
numpy>=1.23
